Fix imports and remove unused functions
